Move origin to bottom of body
Previously, it was on top of the body. This seemed simpler at first,
since it was easily visible, but now that we're moving both the origin
and feet positions dynamically, it's actually easier if the origin is on
the bottom, so the initial body position can just be 0, 0, 0, and the
initial feet altitudes can be 0.